Reconstructing connected components
Most of the code works as before just by renaming things.

Tried to redo the proof that adding an edge again gives a decomposition.
Proved one direction, but for the other got stuck trying to prove:
If you add an edge (x,y) and i and j are now connected, then either:
- they were already connected or
- i was connected with x and j was connected with y or
- j was connected with y and i was connected with x

If we can show this, the proof will follow (did one case using this onproved lemma,
the others should be similar).
The issue is, otherwise there is not way to relate connectedness in the new
graph with connectedness in the original one.