New difinition of path with two non-trivial constructors
One that adds an edge to the left and one to the right.

Also define some functions for concatenating and inverting paths.
Currently have to use the tactic monad to define these because I need induction.
TODO: Figure out how to use structural pattern matching to define these functions.

Tried to define the set of vertices and edges of a path. The problem with defining it
as a set is that function that depend on these sets become uncomputable.
Should probably change the implementation to BST's to get computability and map into bool.

Then we can try to prove Menger's theorem.