Proof that our witness represents connected components.
To show that two vertices u,v ∈ V(G) with c(u) = c(v) are connected
we need a lemma about induction on ℕ, described here:
https://proofassistants.stackexchange.com/questions/1049/strong-induction-on-%e2%84%95-with-function-%ce%b1-%e2%86%92-%e2%84%95
then we use our hypothesis that says that given a vertex v that is not
a root, we can always find and neighbor u with a lower height, then
by induction u is connected to the root, hence by transitivity v is connected
to the root.

The other direction is just induction on the connected relation.

Next part is constructing these witnesses.