Example of a witness for components
A manually constructed example for the graph hog2, that shows it
has 1 connected component.
We define the relvant functions (c, h, root, next) by hand, by cases.

It was quite a pain to construct and `bool_reflect` didn't work in
certain places as one would hope, so I had to define helper lemmas
to actually compute the edge function between certain vertices.

The issue seems like it had something to do with the decidability
of the edge function.